html/user-activation/activation-trigger-keypress.html

<!DOCTYPE html>
<html>
<head>
  <script src="/resources/testharness.js"></script>
  <script src="/resources/testharnessreport.js"></script>
  <script src="/resources/testdriver.js"></script>
  <script src="/resources/testdriver-vendor.js"></script>
  <script src="resources/utils.js"></script>
</head>
<body>
  <h1>Test for keypress activation trigger</h1>
  <p>Tests that a popup is allowed with user activation from a keypress event.</p>
  <input type="text" autofocus />
  <ol id="instructions">
    <li>Press ENTER key.
  </ol>
  <script>
    promise_test(async () => {
        const ENTER_KEY = '\uE007';
        test_driver.send_keys(document.body, ENTER_KEY);
        waitForEvent('keypress')
            .then(consumeTransientActivation())
            .then(consumed => {
                assert_true(consumed,
                            "ENTER keypress event should result in activation");
                return waitForEvent('keyup');
            })
            .then(consumeTransientActivation())
            .then(consumed => {
                assert_false(consumed,
                             "ENTER keyup should have no activation after keypress consumption");
            });
    }, "Activation through keyboard event");
  </script>
</body>
</html>
